Add -O1 optimize flag and fix resulting compiler and linker errors
