+++ .gitignore
@@ -1,3 +1,2 @@
-obj/*
 quokka
 *.swp
