+++ src/attack.h
@@ -7,4 +7,9 @@
 bool square_attacked(Position& pos, Square square, Color side);
 bool in_check(Position& pos);
 
+// Test if a 120 index is on the legal board
+inline bool square_on_board(Square s) {
+	return (to64(s) != OFFBOARD) ? true : false;
+}
+
 #endif
