+++ src/position.h
@@ -6,6 +6,11 @@
 
 using namespace std;
 
+extern Square sq120_to_64[120];
+extern Square sq64_to_120[120];
+
+extern PieceType piece_type[13];
+
 class Position {
 
 public:
@@ -46,9 +51,49 @@ private:
 
 };
 
-Piece create_piece(Color side, PieceType ptype);
+inline Piece create_piece(Color side, PieceType ptype) {
+	return (6 * side + ptype);
+}
 
 extern void parse_UCI_move(Position& pos, string& str);
-extern bool square_on_board(Square s);
 
+// Takes a file and rank and returns their 120 based square index
+inline Square FR2SQ(File f, Rank r) {
+	return ((f + 21) + (r * 10));
+}
+
+// Converts a 120 based square to a 64 based square index
+inline Square to64(Square s) {
+	return sq120_to_64[s];
+}
+
+// Converts a 64 based square to a 120 based square index
+inline Square to120(Square s) {
+	return sq64_to_120[s];
+}
+
+// Convert a 64 based square to a file
+inline File file_of(Square s) {
+	return s & 7;
+}
+
+// Convert a 64 based square to a rank
+inline Rank rank_of(Square s) {
+	return s >> 3;
+}
+
+// Get the color of a piece
+inline Color color_of(Piece p) {
+	return (p <= W_KING) ? WHITE : BLACK;
+}
+
+// Get the type of a piece
+inline PieceType type_of(Piece p) {
+	return piece_type[p];
+}
+
+// Set a bit to zero (for disabling castling perms)
+inline void clear_bit(Byte& i, int bit) {
+	i &= ~(bit);
+}
 #endif // !__POSITION_H__
